Set up wizard modularization
