+++ src/appSettings.js
@@ -1,0 +1,7 @@
+const settings = {
+	parallel : 4,
+	pages : [],
+	pathToSavedPages : './src/pages.csv'
+};
+
+module.exports = settings;
