+++ src/wizzard/welcomeUser.js
@@ -1,0 +1,26 @@
+const strings = require('../strings.json');
+const inquirer = require('inquirer');
+const askForPreferences = require('./askForPreferences');
+
+const crawlFacebook = require('../crawlFacebook');
+
+const settings = require('../appSettings');
+
+
+const welcomeUser = () => {
+	inquirer.prompt([{
+		type: 'list',
+		name: 'welcome',
+		message: strings.english.welcomeWizzard.welcome.question,
+		choices: strings.english.welcomeWizzard.welcome.answers
+	}])
+		.then(answers => {
+			if (answers.welcome === strings.english.welcomeWizzard.welcome.answers[1]) {
+				askForPreferences();
+			} else {
+				crawlFacebook(settings.pages, settings.parallel);
+			}
+		});
+};
+
+module.exports = welcomeUser;
